Publish PointCloud messages from 3D LiDAR
