+++ webots_ros2_universal_robot/setup.py
@@ -10,9 +10,11 @@ worlds = [
     'worlds/universal_robot_multiple.wbt',
     'worlds/universal_robot_rviz.wbt',
     'worlds/universal_robot.wbt',
+    'worlds/universal_robot_lidar.wbt',
     'worlds/.universal_robot_multiple.wbproj',
     'worlds/.universal_robot_rviz.wbproj',
-    'worlds/.universal_robot.wbproj'
+    'worlds/.universal_robot.wbproj',
+    'worlds/.universal_robot_lidar.wbproj'
 ]
 textures = []
 for rootPath, dirNames, fileNames in os.walk('worlds/textures'):
