+++ webots_ros2_universal_robot/worlds/.universal_robot_lidar.wbproj
@@ -1,0 +1,10 @@
+Webots Project File version R2020b
+perspectives: 000000ff00000000fd000000030000000000000263000003e7fc0100000002fb0000001e00480074006d006c0052006f0062006f007400570069006e0064006f00770000000000000002630000006900fffffffb0000001a0044006f00630075006d0065006e0074006100740069006f006e0000000000ffffffff0000006900ffffff000000010000011c0000030cfc0200000001fb0000001400540065007800740045006400690074006f007200000000160000030c0000003f00ffffff0000000300000780000000d9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000007800000006900ffffff000007800000030c00000001000000020000000100000008fc00000000
+simulationViewPerspectives: 000000ff000000010000000200000100000005620100000002010000000100
+sceneTreePerspectives: 000000ff0000000100000002000000c0000000fa0100000002010000000200
+maximizedDockId: -1
+centralWidgetVisible: 1
+orthographicViewHeight: 1
+textFiles: -1
+globalOptionalRendering: LidarPointClouds
+consoles: Console:All:All
